Move some auxiliary test functions into `_zenodo_ome_zarrs.py` module
